# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EDCS API",
        "description": "The Elite Dangerous Companion Suite API provides galaxy data from Elite Dangerous to various consumers, including the EDCS frontend at https://edcs.app. Data is sourced from EDSM (Elite Dangerous Star Map), EDDN (Elite Dangerous Data Network), and the Frontier Companion API (CAPI).",
        "contact": {
            "name": "EDCS",
            "url": "https://edcs.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "EDCS API Server"
        }
    ],
    "paths": {
...
        "/systems/search/distance": {
            "get": {
                "tags": [
                    "System Search"
                ],
                "summary": "Find systems within a given distance of a position",
                "description": "Returns a paginated list of systems within the specified number of light years from a position, sorted by distance. The position can be specified either by a system slug or by raw galactic (x, y, z) coordinates — slug takes precedence if both are provided.",
                "operationId": "4704a33b992c956849d3e6a5fdd5e600",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "System slug ({id64}-{name}) to use as the search origin",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "10477373803-sol"
                        }
                    },
                    {
                        "name": "x",
                        "in": "query",
                        "description": "Galactic X coordinate (required when slug is not provided)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 0
                        }
                    },
                    {
                        "name": "y",
                        "in": "query",
                        "description": "Galactic Y coordinate (required when slug is not provided)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 0
                        }
                    },
                    {
                        "name": "z",
                        "in": "query",
                        "description": "Galactic Z coordinate (required when slug is not provided)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 0
                        }
                    },
                    {
                        "name": "ly",
                        "in": "query",
                        "description": "Search radius in light years (default: 100, max: 5000)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 100
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Maximum results per page (max: 1000)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Systems within the given distance, each with a calculated distance field",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SystemDistance"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/systems/search/information": {
            "get": {
                "tags": [
                    "System Search"
                ],
                "summary": "Search systems by political and demographic attributes",
                "description": "Filters systems by population (minimum), security, government, allegiance, and economy. All text filters are partial-match.",
                "operationId": "702e80bbd475fe8c74593e1f74b1299d",
                "parameters": [
                    {
                        "name": "population",
                        "in": "query",
                        "description": "Minimum population",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 5000000000
                        }
                    },
                    {
                        "name": "security",
                        "in": "query",
                        "description": "Security level (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "high"
                        }
                    },
                    {
                        "name": "government",
                        "in": "query",
                        "description": "Government type (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Democracy"
                        }
                    },
                    {
                        "name": "allegiance",
                        "in": "query",
                        "description": "Allegiance (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Federation"
                        }
                    },
                    {
                        "name": "economy",
                        "in": "query",
                        "description": "Economy type (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Industrial"
                        }
                    },
                    {
                        "name": "withInformation",
                        "in": "query",
                        "description": "Embed political/demographic information",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "withBodies",
                        "in": "query",
                        "description": "Embed celestial bodies",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "withStations",
                        "in": "query",
                        "description": "Embed stations and outposts",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of matching systems",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/System"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SystemBody": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "body_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "name": {
                        "type": "string",
                        "example": "Earth"
                    },
                    "type": {
                        "type": "string",
                        "example": "Planet"
                    },
                    "sub_type": {
                        "type": "string",
                        "example": "Earth-like world"
                    },
                    "distance_to_arrival": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "is_main_star": {
                        "type": "boolean",
                        "example": false
                    },
                    "is_scoopable": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "spectral_class": {
                        "type": "string",
                        "example": "G",
                        "nullable": true
                    },
                    "luminosity": {
                        "type": "string",
                        "example": "Va",
                        "nullable": true
                    },
                    "solar_masses": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "solar_radius": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "absolute_magnitude": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "discovery": {
                        "properties": {
                            "commander": {
                                "type": "string",
                                "example": "CMDR Explorer",
                                "nullable": true
                            },
                            "date": {
                                "type": "string",
                                "example": "3302-06-01",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "radius": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "gravity": {
                        "type": "number",
                        "format": "float",
                        "example": 1,
                        "nullable": true
                    },
                    "earth_masses": {
                        "type": "number",
                        "format": "float",
                        "example": 1,
                        "nullable": true
                    },
                    "surface_temp": {
                        "type": "number",
                        "format": "float",
                        "example": 288,
                        "nullable": true
                    },
                    "is_landable": {
                        "type": "boolean",
                        "example": false
                    },
                    "atmosphere_type": {
                        "type": "string",
                        "example": "Suitable for water-based life",
                        "nullable": true
                    },
                    "volcanism_type": {
                        "type": "string",
                        "nullable": true
                    },
                    "terraforming_state": {
                        "type": "string",
                        "example": "Already terraformed",
                        "nullable": true
                    },
                    "axial": {
                        "properties": {
                            "axial_tilt": {
                                "type": "number",
                                "format": "float",
                                "nullable": true
                            },
                            "semi_major_axis": {
                                "type": "number",
                                "format": "float",
                                "nullable": true
                            },
                            "rotational_period": {
                                "type": "number",
                                "format": "float",
                                "nullable": true
                            },
                            "is_tidally_locked": {
                                "type": "boolean",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "orbital": {
                        "properties": {
                            "orbital_period": {
                                "type": "number",
                                "format": "float",
                                "nullable": true
                            },
                            "orbital_eccentricity": {
                                "type": "number",
                                "format": "float",
                                "nullable": true
                            },
                            "orbital_inclination": {
                                "type": "number",
                                "format": "float",
                                "nullable": true
                            },
                            "arg_of_periapsis": {
                                "type": "number",
                                "format": "float",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "rings": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "parents": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "slug": {
                        "type": "string",
                        "example": "108086401534265707-earth"
                    },
                    "system": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/System"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SystemInformation": {
                "properties": {
                    "allegiance": {
                        "type": "string",
                        "example": "Federation"
                    },
                    "government": {
                        "type": "string",
                        "example": "Democracy"
                    },
                    "population": {
                        "type": "integer",
                        "example": 22780000000
                    },
                    "security": {
                        "type": "string",
                        "example": "High"
                    },
                    "economy": {
                        "type": "string",
                        "example": "Refinery"
                    },
                    "controlling_faction": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "Mother Gaia"
                            },
                            "state": {
                                "type": "string",
                                "example": "None"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "SystemCoords": {
                "properties": {
                    "x": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "y": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "z": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    }
                },
                "type": "object"
            },
            "System": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "id64": {
                        "type": "integer",
                        "example": 10477373803
                    },
                    "name": {
                        "type": "string",
                        "example": "Sol"
                    },
                    "coords": {
                        "$ref": "#/components/schemas/SystemCoords"
                    },
                    "slug": {
                        "type": "string",
                        "example": "10477373803-sol"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "information": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/SystemInformation"
                            }
                        ],
                        "nullable": true
                    },
                    "bodies": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SystemBody"
                        },
                        "nullable": true
                    },
                    "stations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Station"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SystemDistance": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "id64": {
                        "type": "integer",
                        "example": 10477373803
                    },
                    "name": {
                        "type": "string",
                        "example": "Sol"
                    },
                    "coords": {
                        "$ref": "#/components/schemas/SystemCoords"
                    },
                    "distance": {
                        "type": "number",
                        "format": "float",
                        "example": 4.38
                    },
                    "slug": {
                        "type": "string",
                        "example": "10477373803-sol"
                    },
                    "information": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/SystemInformation"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum bearer token. Obtain via POST /auth/login or the Frontier SSO flow.",
                "scheme": "bearer",
                "bearerFormat": "Token"
            }
        }
    },
...
}